+++ Archipelago/src/polygon.hpp
@@ -9,10 +9,10 @@
 #include "shape.hpp"
 #include "geometry.hpp"
 
-double DistancePoint2Point(Coord2D P, Coord2D Q);
-double DistancePoint2Segment(Coord2D P, Coord2D A, Coord2D B);
-double DistancePoint2Ray(Coord2D P, Coord2D A, Coord2D B);
-double DistancePoint2Line(Coord2D P, Coord2D A, Coord2D B);
+float DistancePoint2Point(Coord2D P, Coord2D Q);
+float DistancePoint2Segment(Coord2D P, Coord2D A, Coord2D B);
+float DistancePoint2Ray(Coord2D P, Coord2D A, Coord2D B);
+float DistancePoint2Line(Coord2D P, Coord2D A, Coord2D B);
 
 class Polygon2D : public Shape2D {
 public:
@@ -20,12 +20,12 @@ public:
     std::vector<Coord2D> vertices;
     size_t n; // last element access of n-gone
 
-    double Distance(const Coord2D& P) override {
+    float Distance(const Coord2D& P) override {
         if(vertices.size() < 2) {
             std::cout<<"\tD: "<< DistancePoint2Point(P, vertices[0]);
             return DistancePoint2Point(P, vertices[0]);
         }
-        double ans{DBL_MAX};
+        float ans{FLT_MAX};
         for(size_t i{0}, j{n}; i <= n; j = i++) {
             ans = std::min(ans, DistancePoint2Segment(P, vertices[i], vertices[j]));
         }
@@ -33,11 +33,11 @@ public:
         return ans;
     }
 
-    double Perimeter(void) override {
+    float Perimeter(void) override {
         if(vertices.size() < 3) // degenerate polygon
             return (vertices[0]-vertices[n]).norm();
 
-        double ans{0};
+        float ans{0};
         for(size_t i{0}, j{n}; i <= n; j = i++) {
             ans += (vertices[i]-vertices[j]).norm();
         }
@@ -45,11 +45,11 @@ public:
     }
 
     // http://geomalgorithms.com/a01-_area.html
-    double Area(void) override {
+    float Area(void) override {
         if(vertices.size() < 3) // degenerate polygon
             return 0.0;
 
-        double ans{vertices[0].x*(vertices[1].y - vertices[n].y)};
+        float ans{vertices[0].x*(vertices[1].y - vertices[n].y)};
         for(int i{2}; i <= n; ++i) {
             ans += vertices[i-1].x*(vertices[i].y - vertices[i-2].y);
         }
@@ -64,16 +64,16 @@ public:
         }
         return ans/vertices.size();
     }
-    double eps = std::numeric_limits<double>::epsilon();
+    float eps = std::numeric_limits<float>::epsilon();
     // https://wrf.ecse.rpi.edu/Research/Short_Notes/pnpoly.html
     bool Contains(const Coord2D& P) {
         if(Distance(P) == 0) // on the edge
             return true;
         //determine polygon extremes
-        double minX = vertices[0].x;
-        double maxX = vertices[0].x;
-        double minY = vertices[0].y;
-        double maxY = vertices[0].y;
+        float minX = vertices[0].x;
+        float maxX = vertices[0].x;
+        float minY = vertices[0].y;
+        float maxY = vertices[0].y;
         for(int i = 1 ; i <= n ; i++) {
             Coord2D q = vertices[i];
             minX = std::min( q.x, minX );
